add Area.java

+++ Area.java
@@ -3,12 +3,12 @@ public class Area {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        float Pi = 3.14;
+        float Pi = 3.14f;
 
         System.out.print("Enter num1: ");
-        int num1 = scanner.nextInt();
+        float num1 = scanner.nextInt();
 
-        int sum = Pi * num1;
+        float sum = Pi * num1;
         System.out.println("Area = " + sum);
 
     }
